+++ bedrock/ceda_usa/extract/allocation/mecs.py
@@ -4,7 +4,7 @@ import posixpath
 import pandas as pd
 
 from bedrock.ceda_usa.utils.gcp import GCS_CEDA_INPUT_DIR
-from bedrock.utils.gcp import load_from_gcs
+from bedrock.utils.io.gcp import load_from_gcs
 
 GCS_MECS_DIR = posixpath.join(GCS_CEDA_INPUT_DIR, "EIA_MECS_2018")
 IN_DIR = os.path.join(os.path.dirname(__file__), "..", "input_data")
