+++ bedrock/ceda_usa/utils/snapshots/loader.py
@@ -5,7 +5,7 @@ import pandas as pd
 
 from bedrock.ceda_usa.utils.gcp import GCS_CEDA_USA_DIR
 from bedrock.ceda_usa.utils.snapshots.names import SnapshotName
-from bedrock.utils.gcp import download_gcs_file_if_not_exists
+from bedrock.utils.io.gcp import download_gcs_file_if_not_exists
 
 SNAPSHOT_BASE = os.path.dirname(__file__)
 GCS_SNAPSHOT_DIR = posixpath.join(GCS_CEDA_USA_DIR, "snapshots")
